+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [realm_management_api] Add health endpoint
 - [housekeeping] Add Prometheus instrumenters and exporters.
 - [trigger_engine] Add health endpoint.
+- [housekeeping] Add health endpoint.
 
 ## [0.11.0-beta.2] - 2020-01-24
 ### Added