+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [housekeeping] Add health endpoint.
 - [realm_management] Add health endpoint.
 - [pairing] Add health endpoint.
+- [data_updater_plant] Add health endpoint.
 
 ## [0.11.0-beta.2] - 2020-01-24
 ### Added